small clean up before rebase
