+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ExploreV2/Scripts/Sections/PlacesAndEventsSection/SubSections/EventsSubSection/EventsSubSectionMenu/EventsSubSectionComponentController.cs
@@ -6,44 +6,6 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-public interface IEventsSubSectionComponentController : IDisposable
-{
-    /// <summary>
-    /// It will be triggered when the sub-section want to request to close the ExploreV2 main menu.
-    /// </summary>
-    event Action OnCloseExploreV2;
-
-    /// <summary>
-    /// Request all events from the API.
-    /// </summary>
-    void RequestAllEvents();
-
-    /// <summary>
-    /// Load the featured events with the last requested ones.
-    /// </summary>
-    void LoadFeaturedEvents();
-
-    /// <summary>
-    /// Load the trending events with the last requested ones.
-    /// </summary>
-    void LoadTrendingEvents();
-
-    /// <summary>
-    /// Load the upcoming events with the last requested ones.
-    /// </summary>
-    void LoadUpcomingEvents();
-
-    /// <summary>
-    /// Increment the number of upcoming events loaded.
-    /// </summary>
-    void ShowMoreUpcomingEvents();
-
-    /// <summary>
-    /// Load the going events with the last requested ones.
-    /// </summary>
-    void LoadGoingEvents();
-}
-
 public class EventsSubSectionComponentController : IEventsSubSectionComponentController
 {
     public event Action OnCloseExploreV2;
@@ -125,16 +87,14 @@ public class EventsSubSectionComponentController : IEventsSubSectionComponentCon
     internal void RequestAllEventsFromAPI()
     {
         eventsAPIApiController.GetAllEvents(
-            OnSuccess: eventList =>
-            {
-                eventsFromAPI = eventList;
-                OnRequestedEventsUpdated();
-            },
+            OnSuccess: OnRequestedEventsUpdated,
             OnFail: error => { Debug.LogError($"Error receiving events from the API: {error}"); });
     }
 
-    private void OnRequestedEventsUpdated()
+    private void OnRequestedEventsUpdated(List<EventFromAPIModel> eventList)
     {
+        eventsFromAPI = eventList;
+
         LoadFeaturedEvents();
         LoadTrendingEvents();
         LoadUpcomingEvents();
