+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ExploreV2/Scripts/Sections/PlacesAndEventsSection/SubSections/PlacesSubSection/PlacesSubSectionMenu/IPlacesSubSectionComponentController.cs
@@ -1,0 +1,27 @@
+﻿// unset:none
+using System;
+using System.Collections.Generic;
+
+public interface IPlacesSubSectionComponentController : IDisposable
+{
+    /// <summary>
+    /// It will be triggered when the sub-section want to request to close the ExploreV2 main menu.
+    /// </summary>
+    event Action OnCloseExploreV2;
+
+    /// <summary>
+    /// Request all places from the API.
+    /// </summary>
+    void RequestAllPlaces();
+
+    /// <summary>
+    /// Load the places with the last requested ones.
+    /// </summary>
+    /// <param name="placeList"></param>
+    void LoadPlaces(List<HotScenesController.HotSceneInfo> placeList);
+
+    /// <summary>
+    /// Increment the number of places loaded.
+    /// </summary>
+    void ShowMorePlaces();
+}
